fix(test): explicit .toLong() for JUnit assertEquals
Kotlin Int does not auto-widen to Java long. Explicitly convert both
expected and actual to Long to match assertEquals(String, long, long).

+++ src/android/app/src/test/java/com/openminis/app/tools/ToolFailureHookTest.kt
@@ -164,7 +164,7 @@ class ToolFailureHookTest {
         val summaryEnd = block.indexOf("\n", summaryStart)
         val summary = block.substring(summaryStart, summaryEnd).trim()
         assertTrue("summary truncated to 120", summary.length <= ToolFailureHook.SUMMARY_MAX_LENGTH)
-        assertEquals("summary should be exactly 120 chars for 300-char input", ToolFailureHook.SUMMARY_MAX_LENGTH, summary.length.toLong())
+        assertEquals("summary should be exactly 120 chars for 300-char input", ToolFailureHook.SUMMARY_MAX_LENGTH.toLong(), summary.length.toLong())
     }
 
     // ─── multi-line output: summary is first line ───────────────────────────
